KPS-64 update: Restrict access to teachers
１．担任教師とそれ以外で表示する画面や実行する処理を変更しました
２．教師に生徒を作成できる権限を追加しました
------------------------------------------------------------------------------------------------------------------
 Changes to be committed:
	modified:   src/app/Http/Controllers/AnnouncementController.php
	modified:   src/app/Http/Controllers/ClassesController.php
	modified:   src/app/Http/Controllers/CurriculumController.php
	modified:   src/app/Http/Controllers/LessonController.php
	modified:   src/app/Http/Controllers/StudentController.php
	modified:   src/app/Http/Controllers/TeacherController.php
	modified:   src/app/Models/Teacher.php
	modified:   src/database/migrations/2022_04_22_024348_create_teachers_table.php
	modified:   src/database/seeders/TeacherSeeder.php
	modified:   src/resources/views/admin/register_teacher.blade.php
	new file:   src/resources/views/teacher/common/detail_announcement.blade.php
	new file:   src/resources/views/teacher/common/show_announcements.blade.php
	new file:   src/resources/views/teacher/common/show_class.blade.php
	new file:   src/resources/views/teacher/common/show_lesson.blade.php
	new file:   src/resources/views/teacher/common/show_students.blade.php
	new file:   src/resources/views/teacher/common/show_teachers.blade.php

+++ src/app/Http/Controllers/AnnouncementController.php
@@ -6,6 +6,7 @@ use App\Models\Announcement;
 use App\Models\Classes;
 use App\Models\Semester;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class AnnouncementController extends Controller
 {
@@ -17,15 +18,28 @@ class AnnouncementController extends Controller
         $announcements = $announcement_model->getAnnouncements($class_id);
         $semester_model = new Semester();
         $semester_name = $semester_model->getSentenceOnClass($class_id);
-        return view('teacher.announcement.show_announcements')->with([
-            'class' => $class,
-            'semester_name' => $semester_name,
-            'announcements' => $announcements,
-        ]);
+        $login_user = Auth::user();
+        // 担任以外の教師は削除ボタンを表示しない
+        if(!$class->isResponsibleClass($login_user->id))
+            return view('teacher.common.show_announcements')->with([
+                'class' => $class,
+                'semester_name' => $semester_name,
+                'announcements' => $announcements,
+            ]);
+        else
+            return view('teacher.announcement.show_announcements')->with([
+                'class' => $class,
+                'semester_name' => $semester_name,
+                'announcements' => $announcements,
+            ]);
     }
 
     public function create($class_id){
+        $login_user = Auth::user();
         $class = Classes::find($class_id);
+        // 担任以外の教師はホームにリダイレクトさせる
+        if(!$class->isResponsibleClass($login_user->id))
+            return redirect()->route('teacher.home');
         $semester_model = new Semester();
         $semester_name = $semester_model->getSentenceOnClass($class_id);
         return view('teacher.announcement.create_announcement')->with([
@@ -35,7 +49,11 @@ class AnnouncementController extends Controller
     }
 
     public function store($class_id, Request $request){
+        $login_user = Auth::user();
         $class = Classes::find($class_id);
+        // 担任以外の教師はホームにリダイレクトさせる
+        if(!$class->isResponsibleClass($login_user->id))
+            return redirect()->route('teacher.home');
         $request->validate([
             'title' => 'required|max:31',
             'content' => 'required|max:511'
@@ -53,14 +71,28 @@ class AnnouncementController extends Controller
         $announcement = Announcement::find($announcement_id);
         $semester_model = new Semester();
         $semester_name = $semester_model->getSentenceOnClass($class_id);
-        return view('teacher.announcement.edit_announcement')->with([
-            'class' => $class,
-            'semester_name' => $semester_name,
-            'announcement' => $announcement,
-        ]);
+        $login_user = Auth::user();
+        // 担任以外の教師は編集させない
+        if(!$class->isResponsibleClass($login_user->id))
+            return view('teacher.common.detail_announcement')->with([
+                'class' => $class,
+                'semester_name' => $semester_name,
+                'announcement' => $announcement,
+            ]);
+        else
+            return view('teacher.announcement.edit_announcement')->with([
+                'class' => $class,
+                'semester_name' => $semester_name,
+                'announcement' => $announcement,
+            ]);
     }
 
     public function update($class_id, $announcement_id, Request $request){
+        $login_user = Auth::user();
+        $class = Classes::find($class_id);
+        // 担任以外の教師はホームにリダイレクトさせる
+        if(!$class->isResponsibleClass($login_user->id))
+            return redirect()->route('teacher.home');
         $announcement = Announcement::find($announcement_id);
         $request->validate([
             'title' => 'required|max:31',
@@ -73,6 +105,11 @@ class AnnouncementController extends Controller
     }
 
     public function destroy($class_id, $announcement_id){
+        $login_user = Auth::user();
+        $class = Classes::find($class_id);
+        // 担任以外の教師はホームにリダイレクトさせる
+        if(!$class->isResponsibleClass($login_user->id))
+            return redirect()->route('teacher.home');
         $announcement = Announcement::find($announcement_id);
         $announcement->delete();
         return redirect()->route('announcement.show_at_teacher', $class_id);
