+++ src/app/Http/Controllers/CurriculumController.php
@@ -2,8 +2,10 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\Classes;
 use App\Models\Curriculum;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Validation\Rule;
 
@@ -11,6 +13,11 @@ class CurriculumController extends Controller
 {
     //１コマ追加処理
     public function store($class_id, Request $request){
+        $login_user = Auth::user();
+        $class = Classes::find($class_id);
+        // 担任以外の教師はホームにリダイレクトさせる
+        if(!$class->isResponsibleClass($login_user->id))
+            return redirect()->route('teacher.home');
         $request->validate([
             'lesson_id' => [
                 'integer',
@@ -37,6 +44,11 @@ class CurriculumController extends Controller
 
     //１コマ編集処理
     public function update($class_id, $curriculum_id, Request $request){
+        $login_user = Auth::user();
+        $class = Classes::find($class_id);
+        // 担任以外の教師はホームにリダイレクトさせる
+        if(!$class->isResponsibleClass($login_user->id))
+            return redirect()->route('teacher.home');
         $request->validate([
             'lesson_id' => 'integer'
         ]);
@@ -48,6 +60,11 @@ class CurriculumController extends Controller
 
     //１コマ削除処理
     public function destroy($class_id, $curriculum_id){
+        $login_user = Auth::user();
+        $class = Classes::find($class_id);
+        // 担任以外の教師はホームにリダイレクトさせる
+        if(!$class->isResponsibleClass($login_user->id))
+            return redirect()->route('teacher.home');
         $curriculum = Curriculum::find($curriculum_id);
         $curriculum->delete();
         return redirect()->route('class.edit', $class_id);
